requirements.txt
streamlit
python-docx
PyPDF2
pdfminer.six
spacy
scikit-learn
nltk 